implemented pool total calculation
- calculated total price that includes the price for the hours reserved plus the fees and taxes (current California state tax). Currently Airbnb charges 14% for a service fee but since pools are a much smaller asset, Swimster will charge 8%.
- changed total data type from BIGINT to DOUBLE PRECISION. This was to return decimal values
- checked if guests the user is bringing is <= the guests that the pool listing allows

+++ swimster-api/swimster-schema.sql
@@ -40,7 +40,7 @@ CREATE TABLE reservations (
     start_time  TIME NOT NULL,
     end_time    TIME NOT NULL,
     guests      INT NOT NULL,
-    total       BIGINT NOT NULL,
+    total       DOUBLE PRECISION NOT NULL,
     status      BOOLEAN NOT NULL DEFAULT FALSE,
     created_at  TIMESTAMP NOT NULL DEFAULT NOW()
 );
